split appconfigcache file
