+++ src/Altinn.App.Api/Extensions/ServiceCollectionExtensions.cs
@@ -11,6 +11,7 @@ using Altinn.App.Api.Infrastructure.Telemetry;
 using Altinn.App.Core.Constants;
 using Altinn.App.Core.Extensions;
 using Altinn.App.Core.Features;
+using Altinn.App.Core.Features.Cache;
 using Altinn.App.Core.Features.Correspondence.Extensions;
 using Altinn.App.Core.Features.Maskinporten;
 using Altinn.App.Core.Features.Maskinporten.Extensions;
@@ -44,6 +45,8 @@ public static class ServiceCollectionExtensions
     /// </summary>
     public static void AddAltinnAppControllersWithViews(this IServiceCollection services)
     {
+        // We add this here because it uses a hosted service and we want it to run as early as possible
+        // so that consumers of the cache can rely on it being available.
         services.AddAppConfigurationCache();
 
         // Add API controllers from Altinn.App.Api
