+++ src/Altinn.App.Core/Features/Cache/AppConfigurationCache.cs
@@ -5,23 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace Altinn.App.Core.Features;
-
-internal static class AppConfigurationCacheDI
-{
-    public static IServiceCollection AddAppConfigurationCache(this IServiceCollection services)
-    {
-        services.AddSingleton<AppConfigurationCache>();
-        services.AddSingleton<IAppConfigurationCache>(sp => sp.GetRequiredService<AppConfigurationCache>());
-        services.AddSingleton<IHostedService>(sp => sp.GetRequiredService<AppConfigurationCache>());
-        return services;
-    }
-}
-
-internal interface IAppConfigurationCache
-{
-    public ApplicationMetadata ApplicationMetadata { get; }
-}
+namespace Altinn.App.Core.Features.Cache;
 
 internal sealed class AppConfigurationCache(
     ILogger<AppConfigurationCache> logger,
