+++ src/Altinn.App.Core/Features/Cache/IAppConfigurationCache.cs
@@ -1,0 +1,8 @@
+using Altinn.App.Core.Models;
+
+namespace Altinn.App.Core.Features.Cache;
+
+internal interface IAppConfigurationCache
+{
+    public ApplicationMetadata ApplicationMetadata { get; }
+}
